Fixed order of updates on scrollbar
The handling of scrollbar position changes had some extraneous
code that was removed and an ordering problem on the notification
processing code inside the scene that handles the GDTerm and the
scrollbar.  You must update the max_value and page BEFORE updating
the value.  Otherwise the value is manipulated "automagically" by
the scrollbar object.