Added support for extending and autoscrolling
When the shift is pressed and it would normally start a new selection
and a selection is present, it will just extend the same selection.

When dragging is being performed and the mouse is either above or below
the window it will scroll (if available that amount of rows up or down)
so that the region can be selected.