Change container to be margincontainer
The panelcontainer classes that the main and term_container were
derived from had style implications that causes too much space to
be surrounding the terminal area in many styles.  This is especially
true when in the bottom area.

Changing to MarginContainer and overriding to 0 margin maximizes the
usable area.

The consequence of this however is that when used in the main panel there
isn't enough of a border.  Instead an additional scene is used which
derives from panelcontainer to use the theme specific margins for this
container when used in the main window.