Change paste to use CR when CRLF in paste buffer
Since the terminal code basically inserts the characters as if
typed, the user will only type CR to generate a new line and not
CRLF.  Windows command prompt window ignores the extra LF but the
built in VI editor in git does not.  This results in multiple lines
being generated.