Improved copy and paste handling
The GDTerm class now supports keyboard control for copy and paste.
In particular a Ctrl-Shift-C will copy the selected text and a
Ctrl-Shift-V will paste text.

In addition when multiple lines of text are selected it will no longer
use the large amounts of white space in the terminal window, but show
and copy only from the first column to the last column with a non-space
glyph.