Support initial commands on term starting
Added support for a editor setting that allows a set of commands to be
executed upon initial startup of a terminal window.  This required pushing
the settings down to the inidividual terminal windows and all children.

To make this work reliably the pseudo-terminal interface needed to not return
from start until both the to and from threads were running.