Use updated libtmt library to force UTF8
The libtmt library has a feature to force the use of UTF8 in the
character stream when interpretting input.  That is what the windows
pseudo terminal is expecting on input and output.  This change
required the use of the libtmt provided UTF8 conversion commands so
that when a tmt_wchar_t is received it can be converted back to UTF8
and then given to the terminal renderer (in our case the Godot based
libraries that can take UTF8 but know nothing of tmt_wchar_t).