Handle unicode combiners
Unicode characters that act as combiners (such as for accents or
in the creation of emojis) these characters are passed to the
Godot font subsystem for rendering.

The system fonts work better than the SourceCodePro fonts originally
incorporated into this plugin, so the fonts have been changed to
use system fonts with a fallback to Noto Color Emoji (which is
expected to be installed on the system).

Unit test code was added to help test the use of the underlying
libtmt library.