Fix support for strcpy_s
The strcpy_s is apparently optional, even though on windows it
says it is deprecated.  It isn't supported by gcc, so switching
back to strcpy.  The build tools were not smart enough to detect
this compile failure and stop, so the scripts are updated to
properly detect a compiler error and propogate that out to the
docker scripts.