Change ctrl handling for slash and tilde
The range used for handling control characters in the key
conversion was incorrect.  The last two characters that get
mapped to control characters are Ctrl-~ and Ctrl-/ and the
godot key numbers for ~ and / are not in the same range as
the other characters.  They had to be mapped specially.