Support Cmd key for Mac and font override
Added support for using the Cmd key for Mac OS users, for copy
and paste.  Also included a font override in the editor settings.
This allows any font to be used and a font size to be chosen.

It is not required for a font to be specified, in which case it
will use the embedded Source Code Pro font.