Address issues with startup and resize
A number of issues existed during startup where the control resizes
multiple times before settling out.  This causes the terminal to
look bad, because there are lines that are stuck in little tiny
column space.

There were also issues with restart of the terminal (at least on
Mac) so it was made to not hang if it couldn't get the processor
status on an exiting process.

Also hardened the handling of the terminal during resize events
so that multiple resizes couldn't happen at the same time and that
it didn't try to take action if the proxy wasn't active yet.